- some cleaning

+++ main.cpp
@@ -8,15 +8,13 @@
 int main(int argc, char *argv[]) {
 
 
-
     std::list<std::shared_ptr<TestCase>> benchmarks;
 
 
     benchmarks.push_back(std::shared_ptr<TestCase>(new LoadAndUnloadBenchmark()));
     benchmarks.push_back(std::shared_ptr<TestCase>(new EnableAndDisableBenchmark()));
 
-
-    int count = 100;
+    unsigned long count = 100;
 
     for (std::shared_ptr<TestCase> testCase: benchmarks) {
         testCase->runTestFully(count);
